feature: added auth pages
