+++ src/components/buttons/PrimaryButton.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+type Props = {
+  text: string;
+};
+
+function PrimaryButton({ text }: Props) {
+  return (
+    <button className=" bg-brand-original text-white p-3 rounded-xl font-medium">
+      {text}
+    </button>
+  );
+}
+
+export default PrimaryButton;
